Revert "Merge branch 'master' of https://github.com/PrefectHQ/ui into permissions-audit"
This reverts commit dd6578181af02590cc77d024c03b9349cfba53fd, reversing
changes made to b2be24e3997d249cb21da82a95815fb324cc7578.
