+++ src/components/Card-Title.vue
@@ -44,10 +44,7 @@ export default {
   <div>
     <v-list-item dense class="px-0" :to="params">
       <v-list-item-avatar class="mr-2" tile>
-        <div v-if="$slots['icon']">
-          <slot name="icon" />
-        </div>
-        <v-icon v-else :color="iconColor" :class="iconClass">
+        <v-icon :color="iconColor" :class="iconClass">
           {{ icon }}
         </v-icon>
       </v-list-item-avatar>
