--- src/graphql/Agent/FlowRuns.gql
@@ -1,11 +1,0 @@
-query AgentSubmittableRuns {
-    flow_run (where: {state: {_eq: "Scheduled"}}) {
-    id
-    name
-    state
-    labels
-    scheduled_start_time
-    version
-    flow_id
-    }
-}
