--- src/graphql/Agent/agent-hooks.gql
@@ -1,12 +1,0 @@
-query AgentHooks{
-  hook (where: {event_type: {_eq: "AgentSLAFailedEvent"}}){
-    event_tags 
-      action {
-        id
-        name
-        action_type
-      }
-    event_type
-    
-  }
-}
