--- src/graphql/Agent/recent-runs.gql
@@ -1,24 +1,0 @@
-query RecentRuns($agentId: uuid, $day: timestamptz) {
-  flow_run(
-     where: { 
-      _and: [
-    {state: { _neq: "Scheduled" }},
-    {agent_id: { _eq: $agentId}},
-      ],
-      _or:[
-    {scheduled_start_time:{ _gt: $day}},   
-    {start_time: { _gt: $day}}
-      ]
-    }
-    order_by: { start_time: desc_nulls_last }
-    limit: 10
-  ) {
-    id
-    name
-    end_time
-    flow_id
-    start_time
-    state
-    agent_id
-  }
-}
