--- src/graphql/Agent/table-flow-runs.gql
@@ -1,30 +1,0 @@
-query TableFlowRuns(
-  $name: String
-  $limit: Int
-  $offset: Int
-  $orderBy: [flow_run_order_by!]
-  $agent_id: uuid
-  $state: [String!]
-) {
-  flow_run(
-    where: {
-      agent_id: {_eq: $agent_id}
-      name: { _ilike: $name }
-      state: { _in: $state, _neq: "Scheduled" }
-    }
-    order_by: $orderBy
-    limit: $limit
-    offset: $offset
-  ) {
-    id
-    name
-    end_time
-    scheduled_start_time
-    start_time
-    state
-    flow {
-      name
-      id
-    }
-  }
-}
