--- src/graphql/Agent/timeline-flow-runs.gql
@@ -1,17 +1,0 @@
-query AgentTimelineFlowRuns($agent_id: uuid) {
-  flow_run(
-    where: {
-       agent_id: { _eq: $agent_id } 
-    }
-    limit: 100
-    order_by: { scheduled_start_time: desc }
-  ) {
-    id
-    name
-    flow_id
-    end_time
-    start_time
-    scheduled_start_time
-    state
-  }
-}
