+++ src/graphql/Dashboard/last-flow-runs.gql
@@ -1,4 +1,4 @@
-query LastFlowRuns($flowId: uuid) {
+query LastFlowRuns($flowId: uuid!) {
   flow_run(
     where: { flow_id: { _eq: $flowId }, state: { _neq: "Scheduled" } }
     order_by: { start_time: desc_nulls_last }
@@ -9,6 +9,5 @@ query LastFlowRuns($flowId: uuid) {
     flow_id
     start_time
     state
-   
   }
 }
